backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
numpy
httpx
python-dotenv
python-multipart
websockets
streamlit
requests
google-generativeai
